Use ints instead of floats for day1

+++ 2019/day1/main_test.go
@@ -4,8 +4,8 @@ import "testing"
 
 func TestFuelNeededByMass(t *testing.T) {
 	tests := []struct {
-		input  float64
-		output float64
+		input  int
+		output int
 	}{
 		{12345, 6147},
 	}
@@ -13,7 +13,7 @@ func TestFuelNeededByMass(t *testing.T) {
 	for _, test := range tests {
 		res := FuelNeededByMass(&test.input)
 		if res != test.output {
-			t.Errorf("Expected %f, got %f", test.output, res)
+			t.Errorf("Expected %d, got %d", test.output, res)
 		}
 	}
 }
